+++ NEWS.md
@@ -10,6 +10,7 @@ Fixes:
 Enhancements:
   * New functionality: expand events and edit the validity of individual detections.
   * Added Section information to the detections and the movement tables.
+  * Automatically convert "-" to NA in the Data column of Thelma log files.
 
 ## actel 1.1.1